% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_GamModel.R
\name{AutodetectGAMTerms}
\alias{AutodetectGAMTerms}
\title{Auto-detect GAM terms}
\usage{
AutodetectGAMTerms(model, hgam = "all")
}
\arguments{
\item{model}{A fitted GAM model object}

\item{hgam}{character describing how to handle hgams; use "d" for the density model, "p" for the prob model, or "b" or "all" for both}
}
\value{
returns a data frame with columns describing the name and type of all model components and relevant smoother parameters
}
\description{
This is a useful function that returns a table of the terms in a GAM, whether they are a one dimensional (linear) term, a two dimensional smooth, or a factor. Also detects the offset.
}
