% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{CrossValidateModel}
\alias{CrossValidateModel}
\title{Cross-validate the model}
\usage{
CrossValidateModel(
  model,
  regmult = 1,
  model.type,
  scale.preds = F,
  data,
  key = NA,
  species = NA,
  folds = 10,
  group = "random"
)
}
\arguments{
\item{model}{a fitted model from either mgcv::gam or maxnet::maxnet}

\item{regmult}{a regularization for maxnet models}

\item{model.type}{character; the type of model being made ("maxnet","cloglog","hgam","gam")}

\item{scale.preds}{should the predictions be scaled (assumes count data for now)}

\item{data}{data frame, presumably the same one used to fit the model}

\item{key}{character; an identifier for each record, such as hauljoin}

\item{species}{character; the species or a column in the data set, optional for gams}

\item{folds}{integer; number of folds if random CV is to be used, otherwise ignored}

\item{group}{character; a column or variable name to be used for pre-defined folds in CV, or "random"}
}
\value{
a list of 4 elements; 1- data frame with observations, predictions, and CV out-of-bag predictions
2- list of models generated for each CV fold
3- scale factor for the original model
4- vector of scale factors used for each CV model
}
\description{
Implement n-fold cross-validation, based on either pre-defined folds or randomly
}
\details{
Works for all models currently in use, including maxnet. For quasi-poisson or negbinom, just use type="gam" and it'll work. Outputs a list with 2 or more elements, so you may need to fish through it to find what you want.
}
