% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Effectsplot.R
\name{Effectsplot}
\alias{Effectsplot}
\title{Effects plot}
\usage{
Effectsplot(
  effects.list,
  region = NA,
  crs = NA,
  nice.names = NULL,
  vars = "all"
)
}
\arguments{
\item{effects.list}{list of data frames describing the effects for each term in a model or ensemble}

\item{region}{character; required region to be used with lon/lat map, must correpond to an akgfmaps baselayer,}

\item{crs}{CRS for lat/lon coordinates if different from akgf}

\item{nice.names}{data frame linking names to nicer version for publication figures}

\item{vars}{character vector with names or list indices to be plotted}
}
\value{
list of ggplot objects containing the individual panels and effects
}
\description{
Improved version of the effects plot that uses ggplot and AKGF to keep the style consistent.
}
