% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{EFHComparison}
\alias{EFHComparison}
\title{EFH map comparison}
\usage{
EFHComparison(old, new, nonEFH = 1, background)
}
\arguments{
\item{old}{raster; map of the original EFH area}

\item{new}{raster; map of a new EFH area to be compared to old}

\item{nonEFH}{integer; the value corresponding to non-EFH}

\item{background}{raster; any raster that can be used to mask the relevant area}
}
\value{
returns a new raster, with coding 1 = non-EFH, 2 = EFH in old but not new;
3 = EFH in new but not old, and 4 = EFH in both
}
\description{
This is a function to quickly make a raster that compares the differences between two EFH rasters. The EFH rasters are usually composed of factor values, so you need to know the meaning of said values.  Usually, 1 is <5\%, 2 is 5-25\%, 3 is 25-50\%, 4 is 50-75\%, and 5 is >75\%.
}
