% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{FindEFHbreaks}
\alias{FindEFHbreaks}
\title{Find EFH breaks}
\usage{
FindEFHbreaks(
  abund.raster,
  method = "cumulative",
  threshold = 0.0513,
  quantiles = c(0.05, 0.25, 0.5, 0.75),
  data = NULL
)
}
\arguments{
\item{abund.raster}{raster; map of predicted abundance}

\item{method}{character; "cumulative" or "percentile" for method of EFH calculation}

\item{threshold}{numeric; a threshold to use with the "percentile" method, default is equivalent to 5\% prob in Poisson}

\item{quantiles}{vector of quantiles (other than 0 and 1, that are desired)}

\item{data}{optional data frame with columns "lat" & "lon" to draw the sample from, instead of entire raster}
}
\value{
vector of breaks for the specified quantiles
}
\description{
Find the break points used for EFH for a given abundance raster, given settings.
}
\details{
This has undergone some recent changes. With the additional of the "sanity check" elsewhere, we no longer recommend supplying the data set. The default threshold of .0513 is the poisson abundance equivalent to a 5\% encounter prob.
}
