% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_GamModel.R
\name{FitGAM}
\alias{FitGAM}
\title{Fit the GAM SDM}
\usage{
FitGAM(
  data,
  gam.formula = NULL,
  reduce = F,
  select = F,
  verbose = F,
  family.gam = "poisson",
  theta = NA,
  link.fx = NA
)
}
\arguments{
\item{data}{A data frame including observations and covariates}

\item{gam.formula}{GAM formula or appropriate character string, should include everything}

\item{reduce}{Logical; should the model selection based on UBRE/GCV be used}

\item{select}{Logical; should the built in process be allowed to remove terms with less than one degree of freedom}

\item{verbose}{Logical; should summary be printed after each iteration}

\item{family.gam}{a family that is accepted by the gam function}

\item{theta}{Numeric; an optional set theta value for nb models}

\item{link.fx}{character; the name of a valid link function}
}
\value{
Returns a fitted GAM model object
}
\description{
Fit a GAM that predicts abundance based on the data. One should  supply a formula directly
}
\details{
If verbose=T, printouts of the model selection process are produced.
}
