% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitHurdleGAM.R
\name{FitHurdleGAM}
\alias{FitHurdleGAM}
\title{Fit the Hurdle GAM}
\usage{
FitHurdleGAM(
  data,
  density.formula = NULL,
  prob.formula = NULL,
  select = F,
  reduce = F,
  verbose = F
)
}
\arguments{
\item{data}{A data frame including observations and covariates}

\item{density.formula}{formula for the abundance component}

\item{prob.formula}{formula for the probability component}

\item{select}{Logical; should the built in process be allowed to remove terms with less than one degree of freedom}

\item{reduce}{Logical; should the model selection based on UBRE/GCV be used}

\item{verbose}{Logical; should a summary be printed after each iteration}
}
\value{
returns a fitted hurdle model object i.e. list of two models
}
\description{
The ziplss models are different enough to need their own function, but otherwise behave the same
}
