% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMaxnet.R
\name{FitMaxnet}
\alias{FitMaxnet}
\title{Fit MaxEnt model}
\usage{
FitMaxnet(data, species, vars, reduce = F, regmult = 1, facs = NULL)
}
\arguments{
\item{data}{a data frame containing the covariates and presence/absence data}

\item{species}{character; the name of the column containing the dependent variable}

\item{vars}{a vector of names for columns that contain the covariates}

\item{reduce}{Logical; removes covariates with 0 influence from the model}

\item{regmult}{Numeric; a  regularization multiplier value}

\item{facs}{a vector of names for columns for covariates that should be treated as factors}
}
\value{
a fitted maxnet model
}
\description{
Fit a MaxEnt model for presence/absence
}
\details{
This function fits a maxnet model. Maxnet is somewhat more finicky about NA values and
covariate names, so code has been added to make sure those match
}
