% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_GamModel.R
\name{GAMStats}
\alias{GAMStats}
\title{GAM Stats}
\usage{
GAMStats(model, data)
}
\arguments{
\item{model}{a GAM model object}

\item{data}{a data frame; usually the same one used to fit the GAM model}
}
\value{
a named vector of decimal values indicating the percent contribution
}
\description{
Make a quick and dirty jackknife estimate of the deviance explained by each variable
}
