% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Ensemble.R
\name{GetEnsembleEffects}
\alias{GetEnsembleEffects}
\title{Get ensemble effects}
\usage{
GetEnsembleEffects(effects.list, model.weights, vars = "all", scale = "log")
}
\arguments{
\item{effects.list}{list of lists of data frame, such as those produced from the GetGAMEffects function}

\item{model.weights}{vector of numeric weights for each model, must match order of effects.list}

\item{vars}{character; vector of desired term names or "all"}

\item{scale}{character; should effects be in "log" or "abund" scale}
}
\value{
list of data frames containing the estimated covariate effects in the ensemble
}
\description{
This function takes a list of effects from individual models, and the model weights, and computes an ensemble effect estimate. If CV models are available, then this is used for the CI.
}
