% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEnsembleVariance.R
\name{GetEnsembleVariance}
\alias{GetEnsembleVariance}
\title{Get ensemble variance}
\usage{
GetEnsembleVariance(model.weights, variance.list, abund.list, ensemble.abund)
}
\arguments{
\item{model.weights}{a vector of model weights}

\item{variance.list}{list of rasters of the variance in model predictions}

\item{abund.list}{list of rasters of the predicted abundance from each model}

\item{ensemble.abund}{raster of ensemble predicted abundance}
}
\value{
raster of ensemble predicted variance in predictions
}
\description{
This function computes the variance in abundance based on a formula from Burnham and Anderson
}
