% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_GamModel.R
\name{GetGAMEffects}
\alias{GetGAMEffects}
\title{Get GAM Effects}
\usage{
GetGAMEffects(
  model,
  data,
  cv.model.list = NULL,
  vars = "all",
  scale = "log",
  scale.factor = 1
)
}
\arguments{
\item{model}{a model fit using the "gam" function from mgcv}

\item{data}{data frame; data used to fit the models and cv models}

\item{cv.model.list}{list;cv models to be used to determine confidence bounds, usually from CrossValidateModel function}

\item{vars}{character; a vector of model terms to be plotted, or "all" to return all terms}

\item{scale}{character; Either "log" or "abundance", what should the outputs be in}

\item{scale.factor}{a factor to multiply results}
}
\value{
a list of data frames containing effect estimates and optionally confidence intervals for each term
}
\description{
Use this version if you want the confidence intervals to be based on the cv runs.
}
