% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Maxent.R
\name{GetMaxnetEffects}
\alias{GetMaxnetEffects}
\title{Get MaxEnt effects}
\usage{
GetMaxnetEffects(
  model,
  data,
  cv.models = NULL,
  vars = "all",
  maxnet2d = NULL,
  add.entropy = T,
  scale.factor = 1,
  scale = "log"
)
}
\arguments{
\item{model}{a maxnet model}

\item{data}{data frame; typically the same data used to fit the model}

\item{cv.models}{list;a list of models from cross validation used for confidence intervals}

\item{vars}{a vector of the names of model terms to be plotted, or "all"}

\item{maxnet2d}{a list of vectors of variables to be treated as 2D; used for consistency with the GAMs}

\item{add.entropy}{logical; should be model entropy be added back to the effects, choose F if not estimating abundance}

\item{scale.factor}{numeric; a scale factor to multiply abundance predictions}

\item{scale}{character; should results be on the scale of the linear predictor"log" or abundance "abund"}
}
\value{
a list of data frames containing the estimated covariate effects and variance measures for those effects
}
\description{
Grab the estimated covariate effects and variances for all the effects.
}
