% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOA_bathy-data.R
\docType{data}
\name{GOA_bathy}
\alias{GOA_bathy}
\title{Bathymetry data from the GOA}
\format{
A raster layer with 12 slots
\describe{
\item{name}{character string with path to raster layer}
\item{datanotation}{Raster file native format}
...
}
}
\source{
Compiled by Jeremy Harris using 1-km resolution data from Mark Zimmerman
}
\usage{
GOA_bathy
}
\description{
A raster layer containing bathymetry covariates for SDMs.
}
\keyword{datasets}
