% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{LoadEFHData}
\alias{LoadEFHData}
\title{Load EFH data}
\usage{
LoadEFHData(
  region,
  raster.path = "//akc0ss-n086/SEA_Programs/RACE_EFH_variables/Variables"
)
}
\arguments{
\item{region}{character; "AI", "GOA", or "EBS"}

\item{raster.path}{character; path to the main folder for the EFH project}
}
\value{
does not return anything, but adds various things to the global environment
}
\description{
Quickly set up the covariate rasters.
}
\details{
Should be called before the Load Map function. It would be a good idea to integrate the two, but I haven't gotten to it. Also, it is unlikely to useful except for this project. Deprecated: Likely to be deleted soon.
}
