% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{LoadMap}
\alias{LoadMap}
\title{Title}
\usage{
LoadMap(
  region,
  parameter.file = "G:/Harris/EFH_copy/Map_Settings.csv",
  covariate.raster = raster.stack,
  coast.file = "//akc0ss-n086/SEA_Programs/RACE_EFH_variables/shapefiles",
  GOA.mask = "//akc0ss-n086/SEA_Programs/RACE_EFH_variables/shapefiles"
)
}
\arguments{
\item{region}{character; "AI", "GOA", or "EBS"}

\item{parameter.file}{a csv file containing various mapping paramters}

\item{covariate.raster}{the raster stack produced by LoadEFHData}

\item{coast.file}{file path for a shapefile of the Alaska landmass, very slow to load}

\item{GOA.mask}{file path an additional shapefile more specific to the GOA survey area}
}
\value{
does not return anything, but adds various things to the global environment
}
\description{
Title
}
