% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_akgfmaps.R
\name{MakeAKGFDensityplot}
\alias{MakeAKGFDensityplot}
\title{Map a continuous variable}
\usage{
MakeAKGFDensityplot(
  region,
  density.map,
  buffer = 1,
  survey.area = "default",
  ext.adjust = "default",
  legend.pos = "default",
  legend.title = NA,
  col.palette = "plasma",
  col.palette.limits = c(0, 1),
  title.name = NA,
  title.pos = "default",
  barheight = 5
)
}
\arguments{
\item{region}{character; which region is being used, must correpond to an akgfmaps baselayer}

\item{density.map}{raster or sf; a map to be plotted}

\item{buffer}{numeric; a value between 0 and 1 that can be used to remove outlying value that mess up the color scale}

\item{survey.area}{sf or raster layer to use, if the desired survey area is different from akgfmaps}

\item{ext.adjust}{vector of length four, representing xmin,xmax, ymin,ymax, to adjust the extent}

\item{legend.pos}{vector of length two with decimal position of legend box}

\item{legend.title}{character; a title for the legend}

\item{col.palette}{character a color palette for}

\item{col.palette.limits}{vector of length two giving start and end points for the color palette}

\item{title.name}{character; a title for the figure, or NA to suppress}

\item{title.pos}{vector of length two with coordinates for for the legend, use NA to suppress}

\item{barheight}{numeric; height of the bar in the legend}
}
\value{
a ggplot object of the map
}
\description{
Makes a pretty good map of any continuous variable
}
