% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_akgfmaps.R
\name{MakeAKGFDotplot}
\alias{MakeAKGFDotplot}
\title{Make dotplot with akfgmaps}
\usage{
MakeAKGFDotplot(
  presence,
  absence = NA,
  highdensity = NA,
  dataCRS = NA,
  region,
  survey.area = "default",
  ext.adjust = "default",
  legend.pos = "default",
  title.name = NA,
  title.count = T,
  title.pos = "default",
  abs.name = "absent",
  pres.name = "present",
  hd.name = "top 10\%",
  abs.col = "steelblue4",
  pres.col = "orange",
  hd.col = "red",
  abs.shape = 16,
  pres.shape = 1,
  hd.shape = 16,
  abs.size = 0.1,
  pres.size = 1,
  hd.size = 1
)
}
\arguments{
\item{presence}{data frame with coordinates (lon/lat) for presence locations}

\item{absence}{data frame with coordinates (lon/lat) for absence locations}

\item{highdensity}{data frame with coordinates (lon/lat) for high density locations}

\item{dataCRS}{CRS for the data, if different from the maps}

\item{region}{character; which region is being used, must correpond to an akgfmaps baselayer}

\item{survey.area}{sf or raster layer to use, if the desired survey area is different from akgfmaps}

\item{ext.adjust}{vector of length four, representing xmin,xmax, ymin,ymax, to adjust the extent}

\item{legend.pos}{vector of length two with a decimal location for the legend, use NA to suppress}

\item{title.name}{character; a title for the figure, or NA to suppress}

\item{title.count}{logical; should a count of the total number of presence locations be added to the title}

\item{title.pos}{vector of length two with coordinates for the title box}

\item{abs.name}{character; a name for the absences in the legend}

\item{pres.name}{character; a name for the presences in the legend}

\item{hd.name}{characer; a name for the high density presences in the legend}

\item{abs.col}{color for absence dots}

\item{pres.col}{color for presence dots}

\item{hd.col}{color for high density dots}

\item{abs.shape}{shape for presence dots}

\item{pres.shape}{shape for absence dots}

\item{hd.shape}{shape for high density dots}

\item{abs.size}{size for absence dots}

\item{pres.size}{size for presence dots}

\item{hd.size}{size for high density dots}
}
\value{
a ggplot with the desired figure
}
\description{
makes a pretty good dotplot; needs additional testing with areas other than "bs.all", "goa", and "ai".
}
\details{
Settings are set to work with 8x8 in 300 res output, so results may vary when making different size figures.
}
