% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_akgfmaps.R
\name{MakeAKGFEFHplot}
\alias{MakeAKGFEFHplot}
\title{Make EFH map with akgfmaps}
\usage{
MakeAKGFEFHplot(
  region,
  efh.map,
  drop = 10^8,
  survey.area = "default",
  ext.adjust = "default",
  legend.pos = "default",
  legend.labels = c("95\%", "75\%", "50\%", "25\%"),
  legend.title = NA,
  col.palette = "viridis",
  col.palette.limits = c(0, 1),
  title.name = NA,
  title.pos = "default"
)
}
\arguments{
\item{region}{character; which region is being used, must correspond to an akgfmaps baselayer}

\item{efh.map}{raster of factors to be interpretted as EFH}

\item{survey.area}{sf or raster layer to use, if the desired survey area is different from akgfmaps}

\item{ext.adjust}{vector of length four, representing xmin,xmax, ymin,ymax, to adjust the extent}

\item{legend.pos}{vector of length two with decimal position of legend box}

\item{legend.labels}{vector of labels for the legend, should match number of factors}

\item{legend.title}{character; a title for the legend}

\item{col.palette}{character a color palette for}

\item{col.palette.limits}{vector of length two giving start and end points for the color palette}

\item{title.name}{character; a title for the figure, or NA to suppress}

\item{title.pos}{vector of length two with coordinates for for the legend, use NA to suppress}
}
\value{
raster of EFH quantiles (subareas)
}
\description{
Create EFH map for a species/lifestage combination
}
