% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCrossValidationPlots.R
\name{MakeCrossValidationPlots}
\alias{MakeCrossValidationPlots}
\title{Make cross-validation plots}
\usage{
MakeCrossValidationPlots(error.data, method = "pearson", make.hist = T)
}
\arguments{
\item{error.data}{data frame containing observation, predictions, and CV predictions}

\item{method}{character; a method to be passed to the cor function}

\item{make.hist}{should the histograms be plotted, they sometimes fail and may need to be turned off}
}
\value{
nothing, but creates some plots
}
\description{
Use the model and the cross validated errors to make some residual plots
}
\details{
This is low priority, but could be much better.
}
