% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Xtable.R
\name{MakeDevianceTable}
\alias{MakeDevianceTable}
\title{Make deviance table}
\usage{
MakeDevianceTable(
  model.names = NULL,
  model.types = NULL,
  nice.names = NULL,
  dev.list,
  model.weights = NULL,
  filename
)
}
\arguments{
\item{model.names}{optional vector of names for the models, should match ordering of dev.list}

\item{model.types}{optional vector of model types, should match ordering of dev.list}

\item{nice.names}{optional data.frame with nicer names matched to abbreviated version used in models}

\item{dev.list}{a list of named vectors giving the deviance explained values for each model}

\item{model.weights}{optional vector of ensemble weights}

\item{filename}{path to files}
}
\value{

}
\description{
Make deviance table
}
