% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Ensemble.R
\name{MakeEnsemble}
\alias{MakeEnsemble}
\title{Make model ensemble}
\usage{
MakeEnsemble(
  ensemble.list = NA,
  rmse = NA,
  names.vec = NA,
  minimum = NA,
  model.types = NULL
)
}
\arguments{
\item{ensemble.list}{list of data frames from which rmse can be calculated, not recommended}

\item{rmse}{named vector of rmse values to calculate weights}

\item{names.vec}{vector of names corresponding to the list or rmse vector}

\item{minimum}{numeric; drops any model given less weight than this value}

\item{model.types}{vector of names of models, only one model of a given type will advance}
}
\value{
a named vector with weights for each model
}
\description{
Make an ensemble model vector with weights for each model
}
\details{
This function calculates the weights for an ensemble model using a very simple rmse method. Future work may allow for more complex methods. One can supply a list of dataset (old way), or one can more sensibly just supply a named vector of RMSE values (better way)
}
