% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeEnsembleAbundance.R
\name{MakeEnsembleAbundance}
\alias{MakeEnsembleAbundance}
\title{Make ensemble abundance}
\usage{
MakeEnsembleAbundance(model.weights, abund.list, filename = "")
}
\arguments{
\item{model.weights}{a vector of model weights}

\item{abund.list}{list of abundance rasters corresponding to model weights}

\item{filename}{character; a file name for the abundance raster}
}
\value{
raster formed from weighted average of abund.list
}
\description{
Make the ensemble into a raster.
}
\details{
The abundance rasters should be a list, and must match the ordering of the weights. After this, use the FindEFHbreaks function and others to generate an EFH map as normal.
}
