% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Xtable.R
\name{MakeEnsembleXtable}
\alias{MakeEnsembleXtable}
\title{Make deviance explained table (for ensemble)}
\usage{
MakeEnsembleXtable(
  model.names = c("maxnet", "cloglog", "hpoisson", "poisson", "negbin"),
  ensemble = T,
  weights = NULL,
  converge.vec = NULL,
  abund.check.vec = NULL,
  cor.method = "pearson",
  preds.table = NULL,
  scale.facs = NULL,
  efh.breaks = NULL,
  areas = NULL,
  filename
)
}
\arguments{
\item{model.names}{vector of names for the models}

\item{ensemble}{logical; is an ensemble included}

\item{weights}{vector of weights for each model other than the ensemble}

\item{converge.vec}{vector of T/F for whether models converged}

\item{abund.check.vec}{vector of T/F for whether models passed the abundance check}

\item{cor.method}{character; a method for calculating the r squared values}

\item{preds.table}{data frame with abundance observations and predictions for each model and the ensemble}

\item{scale.facs}{vector of scale factors for each model other than the ensemble}

\item{efh.breaks}{vector of efh breaks for each model including the ensemble}

\item{areas}{vector of areas for each model, including the ensemble}

\item{filename}{character; filename to write the html table to}
}
\value{
does not return anything, but instead writes a html table at filename
}
\description{
Make an HTML table of deviance explained for a model ensemble.
}
\details{
For now, you need to include each of the elements for the ensemble table to work, so no shortcuts. It is kind of cumbersome and unlikely to be useful for others. May soon be deprecated.
}
