% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_GamModel.R
\name{MakeGAMAbundance}
\alias{MakeGAMAbundance}
\title{Make GAM abundance}
\usage{
MakeGAMAbundance(
  model,
  r.stack,
  scale.factor = 1,
  filename = "",
  land = NULL,
  mask = NULL
)
}
\arguments{
\item{model}{a mgcv GAM model object}

\item{r.stack}{raster stack with all the necessary covariates}

\item{scale.factor}{numeric; scale factor that will multiply the model predictions}

\item{filename}{filename for the output}

\item{land}{raster; a mask to be applied to the output usually landmasses}

\item{mask}{raster; an additional mask to apply to the output, if needed}
}
\value{
raster; usually representing abundance
}
\description{
This function makes abundance rasters for the GAMs. When an offset is present in the model, generates a map for the mean value of the offset
}
