% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Maxent.R
\name{MakeMaxEntAbundance}
\alias{MakeMaxEntAbundance}
\title{Make abundance/distribution raster from MaxEnt model}
\usage{
MakeMaxEntAbundance(
  model,
  maxent.stack,
  scale.fac = 1,
  land = NULL,
  mask = NULL,
  type = "cloglog",
  clamp = F,
  filename = ""
)
}
\arguments{
\item{model}{a fitted maxnet model}

\item{maxent.stack}{a raster stack containing all covariates}

\item{scale.fac}{numeric; scale factor that will multiply the model predictions}

\item{land}{raster; a mask to be applied to the output usually landmasses}

\item{mask}{raster; an additional mask to apply to the output, if needed}

\item{type}{character; use "maxnet" for prob suitable habitat or "cloglog" for approximate abundance}

\item{clamp}{Logical; just leave this as F; shoudl the covariates be restricted to the range observe in fitting the model?}

\item{filename}{a filename to save results, "" writes to active memory}
}
\value{
a raster map with the desired prediction
}
\description{
Make an abundance/distribution raster from any dismo or maxnet model. Depending on the settings, it will apply various masks and the cloglog transformation.
}
