% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeVarianceRasters.R
\name{MakeVarianceRasters}
\alias{MakeVarianceRasters}
\title{Make variance rasters}
\usage{
MakeVarianceRasters(
  model.list,
  raster.stack,
  model.type,
  scale.factor = 1,
  efh.break = NA
)
}
\arguments{
\item{model.list}{list of models produced by the CV folds}

\item{raster.stack}{raster stack of the covariates used for the model}

\item{model.type}{character; the type of model ("maxnet","cloglog","hgam","gam")}

\item{scale.factor}{numeric; a scale factor to be applied}

\item{efh.break}{numeric; the EFH breakpoint for the full model, optionally creates an extra map}
}
\value{
raster of estimate non-parametric variance in model predictions
}
\description{
This function makes a non-parametric estimate of spatial variance based on the cross validation models. In order to do this, it needs to hold a lot of data in memory at once, so this can take awhile.
}
