% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Xtable.R
\name{MakeXtable}
\alias{MakeXtable}
\title{Make deviance explained table with constituent model statistics}
\usage{
MakeXtable(
  model,
  rsq.method = "pearson",
  model.type,
  regmult = 1,
  scale.factor = 1,
  efh.break = NULL,
  nice.names = NULL,
  filename,
  abund = NULL,
  train = NULL,
  test = NULL,
  group = NULL,
  forecast = NULL,
  ncoefs = NULL,
  devs = NULL,
  area = NULL
)
}
\arguments{
\item{model}{a fitted model object produced by mgcv::gam() or maxnet::maxnet()}

\item{rsq.method}{character; a method for the cor() function}

\item{model.type}{character; the type of model used, does not autodetect options: maxnet, cloglog, hgam, gam}

\item{regmult}{numeric; a regularization constant for maxnet models, ignored otherwise}

\item{scale.factor}{numeric; a scale factor for abundance predictions}

\item{efh.break}{numeric; a breakpoint or threshold for EFH area}

\item{nice.names}{data frame; table linking abbreviated covariate names to nicer versions}

\item{filename}{character; a file name for the html table produced}

\item{abund}{vector of abundance observations}

\item{train}{vector of predictions from the model trained on the full data set}

\item{test}{vector of out-of-bag predictions}

\item{group}{vector of group labels}

\item{forecast}{vector of predictions from a third group (not used much anymore)}

\item{ncoefs}{vector of coefficients used in maxnet models}

\item{devs}{vector of deviance explained estimates}

\item{area}{numeric; total EFH area}
}
\value{
no return value; writes an html table at filename
}
\description{
Make a table of deviance explained for a model object.
}
\details{
This is a slightly over-complicated function, but it automatically adjusts the table to whatever data is supplied to it. Basically, first it makes a lot of notes as to where things should go, then it takes a second pass construct the table and put everything in the right place. Known issue: does not calculate the degrees of freedom for factors.
}
