% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxnetCoefs.R
\name{MaxnetCoefs}
\alias{MaxnetCoefs}
\title{Get MaxEnt coefficients}
\usage{
MaxnetCoefs(model, maxnet2d = NULL)
}
\arguments{
\item{model}{a fitted maxnet model}

\item{maxnet2d}{a list of vectors of variables to be treated as 2D; used for consistency with the GAMs}
}
\value{
a named vector listing the number of coefficients that apply to each term
}
\description{
Get the number of coefficients for each variable in a maxnet model
}
\details{
If you want things to match the gams, supply a list for maxnet2d
}
