% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Maxent.R
\name{MaxnetStats}
\alias{MaxnetStats}
\title{Get MaxEnt stats}
\usage{
MaxnetStats(model, maxnet2d = NULL, regmult = 1, data, species)
}
\arguments{
\item{model}{a fitted maxnet model}

\item{maxnet2d}{a list of vectors of variables to be treated as 2D; used for consistency with the GAMs}

\item{regmult}{numeric; a regularization penalty multiplier}

\item{data}{data frame; typically the same data used to fit the model}

\item{species}{character; the name of the column for the dependent variable in the model}
}
\value{
a named vector with estimates of the relative deviance explained by each term
}
\description{
This function institutes a jackknife estimate of the importance of each covariate for a maxnet model. Can sometimes run slowly if there are convergence issues.
}
