% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDE.R
\name{PDE}
\alias{PDE}
\title{Calculate Poisson deviance explained (PDE)}
\usage{
PDE(obs, pred)
}
\arguments{
\item{obs}{vector of observations}

\item{pred}{vector of predictions}
}
\value{
returns the estimate percent deviance explained, assuming a Poisson distribution
}
\description{
A quick function to calculate deviance explained
}
\examples{
mods1 <- c(29, 25, 21, 22)
mods2 <- c(3, 23, 21, 23)
PDE(mods1, mods2)
}
