% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{plotAbundance}
\alias{plotAbundance}
\title{Plot abundance (deprecated)}
\usage{
plotAbundance(
  map,
  map.ext = NULL,
  outline = F,
  outline.lwd = 0.5,
  col.vec = plasma(255),
  legend.name = "Abundance (count)",
  legend.text = 0.65,
  legend = T,
  title = "",
  zmin = 0,
  zquant = 0.98,
  zmax = NA,
  center.scale = F,
  back.col = NA,
  ylab = "Latitude",
  xlab = "Longitude",
  label.size = 1
)
}
\arguments{
\item{label.size}{}
}
\value{

}
\description{
This function primarily provides abundance or CPUE-type maps. However, it does a decent job with any raster that is on a continuous scale, and is highly customizable. Deprecated: Likely to be deleted soon in favor of the AKGF version
}
