% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{plotComparison}
\alias{plotComparison}
\title{Plot EFH comparison}
\usage{
plotComparison(
  map,
  map.ext = NULL,
  col.vec = c("grey85", "#D6556DFF", "#43DF71FF", "#440154C1"),
  title = "",
  ylab = "Latitude",
  xlab = "Longitude"
)
}
\arguments{
\item{map}{raster; map of factor values, usually produced by EFHComparison}

\item{map.ext}{an extent that can limit the plot area}

\item{col.vec}{vector of colors to be used in the map}

\item{title}{character; a main title}

\item{ylab}{character; a label for the y axis}

\item{xlab}{character; a label for the x axis}
}
\value{
nothing; but creates a plot
}
\description{
Deprecated: Likely to be deleted soon in favor of AKGF version - a function to plot the comparison map produced by the above function
}
