% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{plotDots}
\alias{plotDots}
\title{Dot plot of sample locations (soon to be deprecated)}
\usage{
plotDots(
  train.data,
  test.data = NA,
  species,
  figure.name = NULL,
  legPosition = "bottomleft",
  background = raster.stack,
  background.file = NA,
  coast = ak.coast,
  back.col = NA,
  train.col = plasma(4)[3],
  test.col = plasma(4)[1],
  point.size = 1,
  text.size = 1,
  legendtext.size = 1,
  axistext.size = 1,
  horiz = F,
  title = NA,
  dlabels = T,
  dlevels = c(100, 200, 500),
  dlabel.size = 1
)
}
\arguments{
\item{dlabel.size}{}
}
\value{

}
\description{
This function plots the sample locations for a species because you need to interleave different pieces, all the parameters for the AddGrid function are needed. Will expand to add background
}
