% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{plotEffects}
\alias{plotEffects}
\title{Plot effects}
\usage{
plotEffects(effects, nice.names = NULL, vars = "all", land = NULL)
}
\arguments{
\item{effects}{list of data frames with the effects, such as from the GetGAMEffects function}

\item{nice.names}{data frame linking abbreviated names to nicer versions to be used in plottingv(optional)}

\item{vars}{vector of which variables to plot, or "all"}

\item{land}{shapefile used to insert landmasses into map of lat/lon effect}
}
\value{
nothing, but creates plots
}
\description{
Plot effects from an effects list
}
\details{
If the cv.models are provided, it will calculate confidence intervals based on that.
}
