% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{plotEFH}
\alias{plotEFH}
\title{Plot EFH (soon to be deprecated)}
\usage{
plotEFH(
  map,
  map.ext = NULL,
  outline = F,
  outline.lwd = 1,
  col.vec = viridis(4),
  title = "",
  background = NULL,
  ylab = "Latitude",
  xlab = "Longitude",
  zlim = NA,
  label.size = 1
)
}
\arguments{
\item{map}{a raster with factor values equivalent to EFH}

\item{map.ext}{an extent object if one wishes to narrow the plot}

\item{outline}{logical; should an outline be drawn around the EFH areas}

\item{outline.lwd}{numeric; the thickness of the outline}

\item{col.vec}{vector of colors to be used in the map}

\item{title}{character; a main title}

\item{background}{a color for non-EFH areas}

\item{ylab}{character; a y axis label}

\item{xlab}{character; a x axis label}

\item{zlim}{a vector of length two, which limits the factors}

\item{label.size}{numeric; the size of the map labels}
}
\value{
does not return anything, but makes a plot
}
\description{
A function that provides a quick method of making EFH maps. Currently replaced with akgfmaps function.
}
