% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEFHComparison.R
\name{PlotEFHComparison}
\alias{PlotEFHComparison}
\title{Plot EFH comparison maps}
\usage{
PlotEFHComparison(
  old = NA,
  new = NA,
  main = "",
  background,
  leg.name = NULL,
  leg.labels,
  region,
  col.vec = NA,
  label.pos = NA,
  leg.pos = NA,
  ext.adjust.x = NA,
  ext.adjust.y = NA,
  nonEFH = 1
)
}
\arguments{
\item{old}{raster of EFH values}

\item{new}{new raster of EFH values}

\item{main}{character; title for the figure}

\item{background}{raster used to define the survey area, often a covariate raster}

\item{leg.name}{character; name for the legend}

\item{leg.labels}{character vector; names to appear in the legend}

\item{region}{character; which region is being used, must correpond to an akgfmaps baselayer}

\item{col.vec}{vector of colors to be used for the categories (usually length 4)}

\item{label.pos}{coordinates for the title box}

\item{leg.pos}{decimal values for placing the legend}

\item{ext.adjust.x}{vector length 2 to adjust the horizontal extent}

\item{ext.adjust.y}{vector length 2 to adjust the vertical extent}

\item{nonEFH}{integer; the value corresponding to non-EFH in the old and new rasters}
}
\value{
ggplot object with the comparison map
}
\description{
This function makes a canned version of the EFH comparison plots. Could use some more work, and is a little less flexible than other functions at current.
}
