% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raster_stack}
\alias{raster_stack}
\title{GOA covariate data}
\format{
A terra PackedSpatRaster with 6 slots
\describe{
\item{lon}{Longitude}
\item{lat}{Latitude}
\item{bdepth}{Bottom depth in meters}
\item{btemp}{Bottom temperature in deg C}
\item{slope}{The rate of change in bathymetry over a defined area, measured as the first derivative of the bathymetric surface, in degrees of incline (Horn 1981, Dolan and Lucieer 2014.)}
\item{sponge}{CPUE of sponge from bottom trawl surveys. Usually, this is transformed into sponge presence-absence when fitting EFH SDMs.}
}
}
\source{
Compiled by Jeremy Harris using 1-km resolution data from Mark Zimmerman
}
\usage{
data(raster_stack)
}
\description{
A terra SpatRaster containing environmental covariates for the GOA,
as used in the 2023 EFH review.
}
\keyword{datasets}
