% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOA_bathy-data.R, R/data.R
\docType{data}
\name{region_data_all}
\alias{region_data_all}
\title{Region data for abundance for all the EFH species included in the 2022 Review}
\format{
A data frame with 15592 rows and 184 variables

A data frame with 15592 rows and 184 variables
}
\usage{
data(region_data_all)

data(region_data_all)
}
\description{
The variables are as follows:

The variables are as follows:
}
\details{
\itemize{
\item hauljoin.
\item region. EFH area (EBS, GOA, or AI)
\item subregion.
\item vessel. Vessel ID from the survey; matches RACEBASE, FOSS, others
\item cruise. Cruise ID number.
\item haul. Haul number.
...
}

\itemize{
\item hauljoin.
\item region. EFH area (EBS, GOA, or AI)
\item subregion.
\item vessel. Vessel ID from the survey; matches RACEBASE, FOSS, others
\item cruise. Cruise ID number.
\item haul. Haul number.
...
}
}
\keyword{datasets}
