% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_data_ebs}
\alias{region_data_ebs}
\title{EBS region data for abundance for EFH species and lifestages included in the 2023 Review}
\format{
a data frame with 15,592 rows and 184 variables (each species/lifestage gets a column):
\describe{
\item{hauljoin}{A numeric value for joining catch data between catch and haul tables.}
\item{region}{EFH area (EBS)}
\item{subregion}{Subregion.}
\item{vessel}{Vessel ID from the survey; matches RACEBASE, FOSS, others}
\item{cruise}{Cruise ID number}
\item{haul}{Haul number}
}
}
\usage{
data(region_data_ebs)
}
\description{
The variables are as follows:
}
\keyword{datasets}
