% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{RMSE}
\alias{RMSE}
\title{Calculate RMSE}
\usage{
RMSE(pred, obs)
}
\arguments{
\item{pred}{vector of predictions}

\item{obs}{vector of observations}
}
\value{
returns RMSE of obs and preds
}
\description{
A quick function to calculate the RMSE.
}
