% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateEnsemble.R
\name{ValidateEnsemble}
\alias{ValidateEnsemble}
\title{Validate ensemble}
\usage{
ValidateEnsemble(
  pred.list,
  model.weights,
  method = "pearson",
  output = T,
  make.plots = T,
  key = NA,
  latlon = T,
  group = NA,
  histogram = F
)
}
\arguments{
\item{pred.list}{list of data frames containing observations and predictions from constituent models}

\item{model.weights}{named vector of weights for each model}

\item{method}{character; method for the cor() function}

\item{output}{logical; should the data frame be returned, or just plots}

\item{make.plots}{logical; should the plots be output, or just the data frame}

\item{key}{character; a column name present in pred.list that works as a case identifier; usually "hauljoin"}

\item{latlon}{logical; should lat and lon be included in the output table, must be present in pred.list}

\item{group}{character; name of a column to be included in the output table representing CV folds}

\item{histogram}{logical; should histograms be plotted (they often don't look nice)}
}
\value{
data frame of ensemble observations vs predictions (pred and cvpred are the same for the ensemble)
}
\description{
This function makes residual plots for the ensemble model and returns a helpful data frame of ensemble predictions
}
