% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{AddGrid}
\alias{AddGrid}
\title{Add grid}
\usage{
AddGrid(
  legPosition = "bottomleft",
  horiz = F,
  legName = "Percentiles",
  legVals = c("95\%", "75\%", "50\%", "25\%"),
  legend.size = 1,
  col.vec = viridis(4),
  depth = F,
  dlevels = c(100, 200, 500),
  dlabels = F,
  dlabel.size = 0.5,
  bathy = raster.stack$bdepth,
  grid = F,
  ext = map.extent,
  map.grid = world.grid,
  grid.col = rgb(0.8, 0.8, 0.8, 0.5),
  xticks = x.ticks,
  yticks = y.ticks,
  xaxis = x.axis,
  yaxis = y.axis,
  axistext.size = 1,
  coast = ak.coast,
  land.col = "grey30",
  border.col = "grey20",
  plot.yaxis = T,
  plot.xaxis = T
)
}
\arguments{
\item{plot.xaxis}{}
}
\value{

}
\description{
Add a map of Alaska and other features to an existing map
}
\details{
This function adds a map of AK and other features to an existing map, including a legend. It has many options for customizing the appearance. The whole "AddGrid" system is old and ought to be removed in favor of the newer akgfmaps plotting, when I get the time. Deprecated: Likely to be deleted soon.
}
