% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Xtable.R
\name{MakeDevianceTable}
\alias{MakeDevianceTable}
\title{Make deviance table}
\usage{
MakeDevianceTable(
  model.names = NULL,
  model.types = NULL,
  nice.names = NULL,
  dev.list,
  model.weights = NULL,
  filename
)
}
\arguments{
\item{filename}{path to files}
}
\value{

}
\description{
Make deviance table
}
