% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Xtable.R
\name{MakeDevianceTable}
\alias{MakeDevianceTable}
\title{Make deviance table}
\usage{
MakeDevianceTable(
  model.names = NULL,
  model.types = NULL,
  nice.names = NULL,
  dev.list,
  model.weights = NULL,
  filename
)
}
\arguments{
\item{model.names}{character vector with names for the models, should match order of dev.list}

\item{model.types}{character vector with types of model, should match order of dev.list}

\item{nice.names}{dataframe matching abbreviated names to nicer versions}

\item{dev.list}{a list of named vectors giving the deviance explained values for each model}

\item{model.weights}{numeric vector with weights for each model, should match order of dev.list}

\item{filename}{a filename to use for saving the output}
}
\value{
does not return anything, but writes two tables at based on the supplied filename
}
\description{
Make an HTML table of deviance explained for a model ensemble.
}
\details{
Makes a simple html table of deviance explained values for models and the ensemble.
}
