% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LoadMap.R
\name{PDE}
\alias{PDE}
\title{Calculate PDE}
\usage{
PDE(obs, pred)
}
\arguments{
\item{obs}{vector of observations}

\item{pred}{vector of predictions}
}
\value{
returns the estimate percent deviance explained, assuming a Poisson distribution
}
\description{
A quick function to calculate the PDE
}
